+++ src/middlewares/ErrorHandling.ts
@@ -12,10 +12,8 @@ export default class {
    * Mount on application.
    */
   public static mount(app: express.Express) {
-    // Load the config.
-    const config = <Config>Object.assign({
-      error_handler: (err: any): void|Promise<void> => {}
-    }, fs.existsSync(`${process.cwd()}/config/config.js`) ? require(`${process.cwd()}/config/config`) : {});
+    // Load options.
+    const options = this.loadOptions();
 
     // Catch 404 and forward to error handler.
     app.use((req: express.Request, res: express.Response, next: express.NextFunction) => {
@@ -28,11 +26,11 @@ export default class {
       console.error(err);
 
       // Call error handling Hook.
-      if (config.error_handler) {
-        if (Types.isAsyncFunction(config.error_handler))
-          await config.error_handler(err);
+      if (options.error_handler) {
+        if (Types.isAsyncFunction(options.error_handler))
+          await options.error_handler(err);
         else
-          config.error_handler(err);
+          options.error_handler(err);
       }
 
       // Set locals, only providing error in development.
@@ -47,5 +45,26 @@ export default class {
         res.render('error');
       }
     });
+  }
+
+
+  /**
+   * Returns the option.
+   * 
+   * @return {Config} option.
+   */
+  private static loadOptions(): Config {
+    // Options with default values set.
+    const defaultOptions: Config = {
+      error_handler: (err: any): void|Promise<void> => {}
+    };
+
+    // If the options file is not found, the default options are returned.
+    const filePath = `${process.cwd()}/config/config`;
+    if (!fs.existsSync(`${filePath}.js`))
+      return defaultOptions;
+
+    // If an options file is found, it returns options that override the default options.
+    return Object.assign(defaultOptions, require(filePath).default||require(filePath));
   }
 }