+++ types/middlewares/CORS.d.ts
@@ -7,4 +7,10 @@ export default class {
      * Mount on application.
      */
     static mount(app: express.Express): undefined;
+    /**
+     * Returns the option.
+     *
+     * @return {Config} option.
+     */
+    private static loadOptions;
 }