+++ types/middlewares/Local.d.ts
@@ -11,4 +11,10 @@ export default class {
      * Mount on application.
      */
     static mount(app: express.Express): void;
+    /**
+     * Returns the option.
+     *
+     * @return {Config} option.
+     */
+    private static loadOptions;
 }