+++ types/routing/Router.d.ts
@@ -10,4 +10,10 @@ export default class {
      * Mount on application.
      */
     static mount(app: express.Express): void;
+    /**
+     * Returns the option.
+     *
+     * @return {Config} option.
+     */
+    private static loadOptions;
 }