+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Changelog
+
+All notable changes to this project will be documented in this file.
+
+## [1.0.1] - 2021-06-02
+### Fixed
+- Fixed a bug that caused an error when reading options defined in ESM's default Export
+
+[1.0.1]: